refactor: 공고 생성, 편집 컴포넌트 분리 및 ENUM타입 API로 호출
